+++ features/fixtures/maze_runner/Assets/Scripts/Main.cs
@@ -12,6 +12,12 @@ public class Command
     public string scenarioName;
 }
 
+[Serializable]
+public class FixtureConfig
+{
+    public string maze_address;
+}
+
 public class Main : MonoBehaviour
 {
 
@@ -25,28 +31,65 @@ public class Main : MonoBehaviour
     private static extern void ClearPersistentData();
 #endif
 
+    private string _fixtureConfigFileName = "/fixture_config.json";
+
     private const string API_KEY = "a35a2a72bd230ac0aa0f52715bbdc6aa";
     private string _mazeHost;
 
+    private const int MAX_CONFIG_GET_TRIES = 15;
+
     public ScenarioRunner ScenarioRunner;
 
-    public void Start()
+    public IEnumerator Start()
     {
-        Debug.Log("Maze Runner app started");
+        Log("Maze Runner app started");
 
-
-        _mazeHost = "http://localhost:9339";
-
-        if (Application.platform == RuntimePlatform.IPhonePlayer ||
-            Application.platform == RuntimePlatform.Android)
-        {
-            _mazeHost = "http://bs-local.com:9339";
-        }
+        yield return GetFixtureConfig();
 
 #if UNITY_STANDALONE_OSX
         PreventCrashPopups();
 #endif
         InvokeRepeating("DoRunNextMazeCommand",0,1);
+    }
+
+    private IEnumerator GetFixtureConfig()
+    {
+        if (Application.platform == RuntimePlatform.Android ||
+            Application.platform == RuntimePlatform.IPhonePlayer)
+        {
+            var numTries = 0;
+            while (numTries < MAX_CONFIG_GET_TRIES)
+            {
+                var configPath = Application.persistentDataPath + _fixtureConfigFileName;
+                if (File.Exists(configPath))
+                {
+                    var configJson = File.ReadAllText(configPath);
+                    Log("Mazerunner got fixture config json: " + configJson);
+                    var config = JsonUtility.FromJson<FixtureConfig>(configJson);
+                    _mazeHost = "http://" + config.maze_address;
+                    break;
+                }
+                else
+                {
+                    numTries++;
+                    Log(string.Format("Maze Runner did not find the config file at path {0}  try number {1}", configPath, numTries));
+                    yield return new WaitForSeconds(1);
+                }
+            }
+        }
+
+        if (string.IsNullOrEmpty(_mazeHost))
+        {
+            Log("Host not set from config file, using hard coded");
+            _mazeHost = "http://localhost:9339";
+
+            if (Application.platform == RuntimePlatform.IPhonePlayer ||
+                Application.platform == RuntimePlatform.Android)
+            {
+                _mazeHost = "http://bs-local.com:9339";
+            }
+        }
+        Log("Mazerunner host set to: " + _mazeHost);
     }
 
     private void DoRunNextMazeCommand()
@@ -57,7 +100,7 @@ public class Main : MonoBehaviour
     IEnumerator RunNextMazeCommand()
     {
         var url = _mazeHost + "/command";
-        Console.WriteLine("RunNextMazeCommand called, requesting command from: {0}", url);
+        Log("Requesting maze command from: " + url);
         using (UnityWebRequest request = UnityWebRequest.Get(url))
         {
             yield return request.SendWebRequest();
@@ -69,23 +112,23 @@ public class Main : MonoBehaviour
                 !request.isNetworkError;
 #endif
 
-            Console.WriteLine("result is " + result);
+            Log("result is " + result);
             if (result)
             {
                 var response = request.downloadHandler?.text;
-                Console.WriteLine("Raw response: " + response);
+                Log("Raw response: " + response);
                 if (response == null || response == "null" || response == "No commands to provide")
                 {
-                    Console.WriteLine("No Maze Runner command to process at present");
+                    Log("No Maze Runner command to process at present");
                 }
                 else
                 { 
                     var command = JsonUtility.FromJson<Command>(response);
                     if (command != null)
                     {
-                        Console.WriteLine("Received Maze Runner command:");
-                        Console.WriteLine("Action: " + command.action);
-                        Console.WriteLine("Scenario: " + command.scenarioName);
+                        Log("Received Maze Runner command:");
+                        Log("Action: " + command.action);
+                        Log("Scenario: " + command.scenarioName);
 
                         if ("clear_cache".Equals(command.action))
                         {
@@ -136,6 +179,11 @@ public class Main : MonoBehaviour
 #if UNITY_IOS
         ClearPersistentData();
 #endif
+    }
+
+    private static void Log(string msg)
+    {
+        Logger.I(msg);
     }
 
 }