+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Persistence/PersistEventReportCallback.cs
@@ -7,7 +7,6 @@ public class PersistEventReportCallback : Scenario
     {
         base.PrepareConfig(apiKey, host);
         Configuration.Context = "Error 2";
-        Configuration.AutoTrackSessions = false;
         Configuration.AddOnSendError((@event) => {
 
             @event.App.BinaryArch = "Persist BinaryArch";