+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchMaxCacheSize.cs
@@ -9,7 +9,6 @@ public class SwitchMaxCacheSize : Scenario
     {
         base.PrepareConfig(apiKey, host);
         Configuration.SwitchCacheMaxSize = 2097155;
-        Configuration.AutoTrackSessions = false;
     }
 
     public override void Run()