+++ src/BugsnagUnity/INativeClient.cs
@@ -104,5 +104,10 @@ namespace BugsnagUnity
 
         bool ShouldAttemptDelivery();
 
+        /// <summary>
+        /// Subscribes the Unity layer for session callbacks. Not many users use session callbacks, so we are only subscribing to the native side if necessary as an optimisation.
+        /// </summary>
+        void RegisterForOnSessionCallbacks();
+
     }
 }