+++ src/BugsnagUnity/Native/Android/NativeClient.cs
@@ -156,6 +156,11 @@ namespace BugsnagUnity
         {
             return true;
         }
+
+        public void RegisterForOnSessionCallbacks()
+        {
+            NativeInterface.RegisterForOnSessionCallbacks();
+        }
     }
 
 }